% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastAUC_function.R
\name{auc}
\alias{auc}
\title{Calculate and compare AUC estimate for clustered data}
\usage{
auc(test_1, test_2 = NULL, status, cluster = NULL)
}
\arguments{
\item{test_1}{the set of test measurments for which the AUC should be
computed}

\item{test_2}{a second set of test measurments for which the AUC is computed
and compared to the first set of test measurements}

\item{status}{a logical vector representing the gold standard
negative/positive status of an observaiton}

\item{cluster}{variable identifying clusters}
}
\value{
a list containing \item{auc}{a vector of AUC estimates}
  \item{var}{the individual variance estimate or covariance matrix}
  \item{test_stat}{the test statistic for comparing the AUC estimates,
  distributed standard normal} \item{p_value}{the p-value associated with the
  AUC test comparison (2-sided)}
}
\description{
This function can be used to calculate and compare AUC estimates. If a single
set of test measurements is provided, the corresponding AUC and variance
values are computed. If two sets of test measurements are provided, a
significance test for a difference in AUC estimates is also computed.
Optionally, the previous can also be computed in the presence of clustered
data per Obuchowski (1997). This function implicity assumes that higher
values of test_1 and test_2 are correlated with the positive gold standard
(status = TRUE)
}
\examples{
n_obs <- 10000
n_clust <- 100

cluster <- sample(x = 1:n_clust, size = n_obs, replace = T)
status <- as.logical(rbinom(n = n_obs, size = 1,  prob = 0.3))
test_1 <- ifelse(status,
                 rnorm(n = n_obs, mean = 15, sd = 20),
                 rnorm(n = n_obs,mean = 10, sd = 20))
test_2 <- ifelse(status,
                 rnorm(n = n_obs, mean = 20, sd = 20),
                 rnorm(n = n_obs,mean = 10, sd = 20))

auc(test_1 = test_1, status = status)

auc(test_1 = test_1, test_2 = test_2, status = status, cluster = cluster)

}
\references{
Obuchowski N.A. (1997) Nonparametric Analysis of clustered ROC
  Curve Data. Biometrics, 53, 567-578.

  DeLong E.R., DeLong D.M, Clarke-Pearson D.L. (1988) Comparing the areas
  under two or more correlated receiver operating characteristic curves: a
  nonparametric approach. Biometrics, 44, 837-845.
}
